moves all created components into their assigned repos
