+++ jaatube-Clone-App/src/App.jsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import reactLogo from "./assets/react.svg";
 import viteLogo from "/vite.svg";
 import "./App.css";
-import "./About";
+import "./components/about/About";
 
 function App() {
   const [count, setCount] = useState(0);
